Note: Please check the commit comments in the previous push first. This is a minor update.
Update: Add support to specify the number of processors to be used for 2D cartesian grids: nproc_1 x nproc_2. Setting one of the dimension to have only one processor essentially enables slab decomposition. TEST_assemble is now updated to show this slab decomposition --- I did not fully check all the details of the result, but at first glance, it seems to be working fine.

To-do: Once we fully check the slab decomposition, it should now be easy to add support to 2D data (r-theta or r-z) - make wrappers that check which glb_sz(i)==1 will determine the direction that is no longer used, and it is trivial to turn on slab decomposition and call exchange / assemble / disassemble accordingly. Initalise does have built-in support for 2D data, but I havn't fully checked it yet.